refactor: monorepo (#3)
