--- .babelrc
@@ -1,11 +1,0 @@
-{
-  "presets": ["next/babel", "@linaria"],
-  "plugins": [
-    [
-      "@babel/plugin-transform-react-jsx",
-      {
-        "runtime": "automatic"
-      }
-    ]
-  ]
-}
