+++ .eslintrc.js
@@ -1,34 +1,3 @@
-module.exports = {
-  env: {
-    browser: true,
-    es2020: true,
-    node: true,
-    jest: true,
-  },
-  extends: [
-    'plugin:react/recommended',
-    'plugin:@typescript-eslint/recommended',
-    'prettier/@typescript-eslint',
-    'prettier/react',
-  ],
-  parser: '@typescript-eslint/parser',
-  parserOptions: {
-    ecmaFeatures: {
-      jsx: true,
-    },
-    ecmaVersion: 11,
-    sourceType: 'module',
-  },
-  plugins: ['react', '@typescript-eslint', 'prettier'],
-  rules: {
-    'prettier/prettier': 'error',
-  },
-  settings: {
-    'import/resolver': {
-      typescript: {},
-    },
-    react: {
-      version: 'detect',
-    },
-  },
-}
+const config = require('@local/eslint-config')
+
+module.exports = config
