--- next.config.js
@@ -1,5 +1,0 @@
-const withLinaria = require('next-linaria')
-
-module.exports = withLinaria({
-  reactStrictMode: true,
-})
