+++ packages/design-system/.babelrc
@@ -1,0 +1,3 @@
+{
+  "plugins": ["babel-plugin-styled-components"]
+}
