+++ packages/design-system/src/button/button.styles.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const ButtonContainer = styled.div`
+  display: flex;
+`
+
+export const ButtonComponent = styled.button``
