+++ packages/design-system/src/button/index.ts
@@ -1,0 +1,2 @@
+export { default } from './button'
+export type { ButtonProps } from './button'
