+++ packages/eslint-config/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@local/eslint-config",
+  "version": "1.0.0",
+  "private": true,
+  "main": ".eslintrc.js",
+  "license": "MIT",
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "4.26.0",
+    "@typescript-eslint/parser": "4.26.0",
+    "eslint": "7.27.0",
+    "eslint-config-prettier": "8.3.0",
+    "eslint-import-resolver-typescript": "2.4.0",
+    "eslint-plugin-import": "2.23.4",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-prettier": "3.4.0",
+    "eslint-plugin-promise": "5.1.0",
+    "prettier": "2.3.0",
+    "tslint-config-prettier": "1.18.0"
+  }
+}
