+++ packages/front/next.config.js
@@ -1,0 +1,10 @@
+/** @type {import('next').NextConfig} */
+
+const withTranspileModules = require('next-transpile-modules')(['@local/design-system'])
+const withLinaria = require('next-linaria')
+
+module.exports = withLinaria(
+  withTranspileModules({
+    reactStrictMode: true,
+  }),
+)
