+++ packages/front/pages/index.tsx
@@ -1,0 +1,13 @@
+import { NextPage } from 'next'
+
+import Button from '@local/design-system'
+
+const Home: NextPage = () => {
+  return (
+    <div>
+      <Button>Teste</Button>
+    </div>
+  )
+}
+
+export default Home
