+++ packages/front/public/favicon.ico
