+++ packages/front/styles/globals.css
