--- pages/index.tsx
@@ -1,15 +1,0 @@
-import { styled } from '@linaria/react'
-import type { NextPage } from 'next'
-
-const Container = styled.div`
-  max-width: 1024px;
-  margin-right: auto;
-  margin-left: auto;
-  background-color: gray;
-`
-
-const Home: NextPage = () => {
-  return <Container>olaria</Container>
-}
-
-export default Home
